Add files via upload
In this project done in advanced java, MVC is implemented by making dao class and working on it. In this, to open any page other than login page such as 'about us' or 'index', you have to be logged in. If you are not logged in you cannot visit the pages. Just to implement this, I tried doing this on 2 HTML pages.